projects/inter_scrap/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.6.0
django-rest==0.8.6
django-sslserver==0.22
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
optional-django==0.3.0
psycopg2==2.8.6
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
